+++ packages/core/test/test-util/test-services/openapi/swagger-yaml-service/api.d.ts
@@ -1,0 +1,14 @@
+import { OpenApiRequestBuilder } from '../../../../../src';
+import { DefaultApi } from './openapi/api';
+import { TestEntity } from './openapi/model';
+export declare const SwaggerYamlServiceApi: {
+    postEntity: (args: {
+        pathParam: string;
+        queryParam?: string;
+    }) => OpenApiRequestBuilder<DefaultApi, "postEntity">;
+    patchEntity: (args: {
+        pathParam: string;
+        body?: TestEntity;
+    }) => OpenApiRequestBuilder<DefaultApi, "patchEntity">;
+};
+//# sourceMappingURL=api.d.ts.map