+++ packages/core/test/test-util/test-services/openapi/test-service/openapi/model/index.d.ts
@@ -1,0 +1,3 @@
+export * from './simple-test-entity';
+export * from './test-entity';
+//# sourceMappingURL=index.d.ts.map