+++ packages/generator/src/generator.ts
@@ -1,6 +1,10 @@
 import { PathLike, readFileSync } from 'fs';
 import { resolve, basename } from 'path';
-import { createLogger, splitInChunks } from '@sap-cloud-sdk/util';
+import {
+  createLogger,
+  splitInChunks,
+  transpileDirectory
+} from '@sap-cloud-sdk/util';
 import { emptyDirSync } from 'fs-extra';
 import {
   Directory,
@@ -81,20 +85,8 @@ export function transpileDirectories(
   directories: Directory[]
 ): Promise<void[]> {
   return Promise.all(
-    directories.map(directory => transpileDirectory(directory))
+    directories.map(directory => transpileDirectory(directory.getPath()))
   );
-}
-
-export function transpileDirectory(dir: Directory): Promise<void> {
-  logger.info(`Transpiling files in the directory: ${dir.getPath()} started.`);
-  return execa('tsc', { cwd: dir.getPath() })
-    .then(() => {
-      logger.info(`Transpiling files in directory: ${dir.getPath()} finished.`);
-    })
-    .catch(err => {
-      logger.error(`Error: Failed to generate js files: ${err}`);
-      process.exit(1);
-    });
 }
 
 export async function generateProject(