+++ test-packages/test-services/openapi/swagger-yaml-service/index.d.ts
@@ -1,0 +1,3 @@
+export * from './openapi/model';
+export * from './api';
+//# sourceMappingURL=index.d.ts.map